ml-service/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
numpy
opencv-python